added voice channels and moved createClass into bot folder
